+++ reactivestate/src/commonMain/kotlin/com/ensody/reactivestate/ErrorEvents.kt
@@ -17,11 +17,13 @@ public suspend fun <E : ErrorEvents> withErrorReporting(
     if (onError == null) {
         withErrorReporting(eventNotifier) { block() }
     } else {
-        withErrorReporting({
-            withErrorReporting(eventNotifier) {
-                onError(it)
-            }
-        }) { block() }
+        withErrorReporting(
+            {
+                withErrorReporting(eventNotifier) {
+                    onError(it)
+                }
+            },
+        ) { block() }
     }
 }